move user logic to services/user.js
