+++ src/pages/userPanel/MyDetails.jsx
@@ -3,7 +3,7 @@ import { useSelector } from 'react-redux';
 import { useQuery, useMutation } from '@apollo/client';
 import { USER_PERSONAL_DETAILS } from 'graphql/queries/user.js';
 import { UPDATE_USER_DETAILS } from 'graphql/mutations/user.js';
-import handleMyDetailsValidation from 'services/users/handleMyDetailsValidation.js';
+import { handleMyDetailsValidation } from 'services/user.js';
 import TextInput from 'components/inputs/TextInput.jsx';
 import SubmitButton from 'components/SubmitButton.jsx';
 import SuccessModal from 'components/modals/SuccessModal.jsx';
