--- src/services/users/handleRegisterValidation.js
@@ -1,33 +1,0 @@
-import regexps from 'data/regexps.js';
-import { validateByRegexp } from 'utils/helpers.js';
-import { VALIDATION_ERROR_MESSAGES } from 'data/errors.js';
-
-const handleRegisterValidation = ({ email, password, avatars }) => {
-  const isEmailValid = validateByRegexp({ regexp: regexps.email, subject: email });
-  const isPasswordValid = validateByRegexp({ regexp: regexps.password, subject: password });
-  const isAvatarValid = validateAvatars({ avatars });
-
-  return {
-    emailError: !isEmailValid && VALIDATION_ERROR_MESSAGES.email,
-    passwordError: !isPasswordValid && VALIDATION_ERROR_MESSAGES.password,
-    avatarError: !isAvatarValid && VALIDATION_ERROR_MESSAGES.avatar,
-    validationStatus: isEmailValid && isPasswordValid && isAvatarValid
-  };
-};
-
-const validateAvatars = ({ avatars }) => {
-  const allowedFileFormats = ['image/png', 'image/svg+xml', 'image/jpeg'];
-  let eachAvatarsHasValidFormat = true;
-
-  avatars.forEach(({ fileType }) => {
-    const hasAllowedFormat = allowedFileFormats.includes(fileType);
-
-    if (!hasAllowedFormat) {
-      eachAvatarsHasValidFormat = false;
-    }
-  });
-
-  return eachAvatarsHasValidFormat;
-};
-
-export default handleRegisterValidation;
