--- src/services/users/sortAvatars.js
@@ -1,6 +1,0 @@
-export const sortByMainField = (a, b) => {
-  if (a.main === 'true') return 1;
-  if (b.main === 'true') return -1;
-
-  return 0;
-};
